Mark all experiments as equilibrium-model (lg) to leave room for temporal (tlg)
Every current experiment uses the equilibrium Logit-Graph, so prefix them all
with 'lg' to distinguish from the new temporal model (future 'tlg'):
- scripts: rename all scripts/{experiments,gic,closedform,diagnostics}/run_*.py
  to run_lg_*.py (30 files).
- Makefile: prefix every experiment target with 'lg-' (lg-sigma-convergence,
  lg-roc-paper, lg-aic-*, lg-convergence-diagnostics, lg-anova-*-robust,
  lg-gic-<dataset>[-closedform], ...) and point each to its run_lg_*.py script.
- README Experiments section + script docstrings + lg_aic_utils helper comment
  updated to the lg- targets; documents the lg-/tlg- naming convention.

Pure rename/rewire: no src/ or test changes; full suite still 330 passed, 2 xfailed.
Temporal-model experiments will land under the 'tlg-' prefix (none yet).
