+++ scripts/closedform/run_lg_human_connectomes_closedform.py
@@ -1,15 +1,15 @@
 #!/usr/bin/env python3
 """Closed-form (moment-matched) baseline estimators vs fixed-interval grid
-search, on Google+ ego networks, alongside a fairly-scored Logit-Graph (LG).
+search, on OASIS-3 human brain connectomes, alongside a fairly-scored LG.
 
-For each gplus ego network in the size window we score, by spectral GIC
-(2*KL + 2*n_params, KL on the 50-bin normalized-Laplacian density, lower=better):
+Human-connectome twin of run_connectomes_closedform.py. The OASIS-3 dataset has
+several parcellation scales (different node counts) with hundreds of subjects
+each; all subjects in a scale share a node count. We sample PER_SCALE subjects
+from each of a few scales (seeded) and score, by spectral GIC
+(2*KL + 2*n_params, KL on the normalized-Laplacian density, lower=better):
 
   * LG            -- best of d in {0,1,2}, each scored by burn-in + ensemble mean
-                     of the spectral density over N_RUNS post-burn-in snapshots
-                     (same averaging convention the baselines get; avoids the
-                     pipeline's best-of-trajectory cherry-pick — see
-                     diag_lg_convergence.py).
+                     of the spectral density over N_RUNS post-burn-in snapshots.
   * ER/BA/WS      -- two ways each:
        grid : fixed interval, GRID_POINTS points, parameter picked by min GIC
        cf   : closed-form moment estimate (no search)
@@ -23,23 +23,27 @@ Closed-form estimators (n nodes, E edges, kbar = 2E/n avg degree):
   KR  d = round(kbar)          (nd even)   (E = nd/2)
   GRG r = sqrt(kbar / (pi*(n-1)))          (E[deg] ~ (n-1) pi r^2, 2-D)
 
-Reproducible: single fixed seed (LG_CF_SEED), BLAS threads pinned to 1, and the
-gplus tarball is auto-extracted if the .edges files are missing. Read-only w.r.t.
-the library; writes only under runs/gplus_closedform/. Findings:
-FINDINGS_gplus_closedform.md.
+Brain graphs are weighted (fiber density); we binarize to the undirected,
+unweighted largest connected component. Dense eigvalsh for n<=500, deterministic
+KPM above. Reproducible: fixed seed (LG_HCF_SEED) drives the subject sampling
+and all generators; BLAS threads pinned to 1. Read-only w.r.t. the library;
+writes only under runs/human_connectomes_closedform/. Findings:
+FINDINGS_human_connectomes_closedform.md.
 
 Env knobs (all optional):
-  LG_CF_SEED (12345)    LG_CF_QUICK (0 -> full; 1 -> smoke on a few small nets)
-  LG_CF_MIN_NODES (50)  LG_CF_MAX_NODES (300)  LG_CF_MAX_NETS (all)
-  LG_CF_N_RUNS (5)      LG_CF_GRID_POINTS (5)
+  LG_HCF_SEED (12345)   LG_HCF_QUICK (0 -> full; 1 -> smoke, one small scale)
+  LG_HCF_SCALES (oasis3 scale1,scale2,scale3 + repeated_10_scale_250)
+  LG_HCF_PER_SCALE (8)  LG_HCF_N_RUNS (5)  LG_HCF_GRID_POINTS (5)
+  LG_HCF_MIN_NODES (20) LG_HCF_MAX_NODES (500)
 
-  make gic-gplus-closedform        full run (~30s, 17 nets)
-  make gic-gplus-closedform-quick  smoke (~5s, a few small nets)
+  make lg-gic-human-connectomes-closedform        full run
+  make lg-gic-human-connectomes-closedform-quick  smoke
 """
 from __future__ import annotations
 
 import math
 import os
+import random
 import sys
 import time
 import warnings
@@ -76,27 +80,29 @@ def _int(env, default):
     return int(raw) if raw is not None else default
 
 
-QUICK = os.environ.get("LG_CF_QUICK", "0") == "1"
-MIN_NODES = _int("LG_CF_MIN_NODES", 50)
-MAX_NODES = _int("LG_CF_MAX_NODES", 120 if QUICK else 300)  # full matches gic-gplus
-MAX_NETS = _int("LG_CF_MAX_NETS", 4 if QUICK else 10_000)   # full = all nets in window
-N_RUNS = _int("LG_CF_N_RUNS", 3 if QUICK else 5)
-GRID_POINTS = _int("LG_CF_GRID_POINTS", 3 if QUICK else 5)
-LG_D_LIST = [0, 1, 2]                       # LG d exploration
-# Fair LG scoring: burn-in then N_RUNS spectral snapshots (stride apart),
-# scaled with n. Burn-in past the noisy walk seen in diag_lg_convergence.py.
-LG_BURN_MIN = _int("LG_CF_LG_BURN_MIN", 4000)
-LG_BURN_PER_N = _int("LG_CF_LG_BURN_PER_N", 25)
-LG_STRIDE_MIN = _int("LG_CF_LG_STRIDE_MIN", 600)
-LG_STRIDE_PER_N = _int("LG_CF_LG_STRIDE_PER_N", 6)
-SEED = _int("LG_CF_SEED", 12345)
+QUICK = os.environ.get("LG_HCF_QUICK", "0") == "1"
+_DEFAULT_SCALES = ("oasis3_graphmls_scale1,oasis3_graphmls_scale2,"
+                   "oasis3_graphmls_scale3,repeated_10_scale_250")
+SCALES = os.environ.get(
+    "LG_HCF_SCALES", "repeated_10_scale_33" if QUICK else _DEFAULT_SCALES
+).split(",")
+PER_SCALE = _int("LG_HCF_PER_SCALE", 3 if QUICK else 8)
+MIN_NODES = _int("LG_HCF_MIN_NODES", 20)
+MAX_NODES = _int("LG_HCF_MAX_NODES", 500)
+N_RUNS = _int("LG_HCF_N_RUNS", 3 if QUICK else 5)
+GRID_POINTS = _int("LG_HCF_GRID_POINTS", 3 if QUICK else 5)
+LG_D_LIST = [0, 1, 2]
+LG_BURN_MIN = _int("LG_HCF_LG_BURN_MIN", 4000)
+LG_BURN_PER_N = _int("LG_HCF_LG_BURN_PER_N", 25)
+LG_STRIDE_MIN = _int("LG_HCF_LG_STRIDE_MIN", 600)
+LG_STRIDE_PER_N = _int("LG_HCF_LG_STRIDE_PER_N", 6)
+SEED = _int("LG_HCF_SEED", 12345)
 
-# Current fixed intervals (mirror simulation.py defaults).
 GRID_INTERVALS = {"ER": (0.01, 0.25), "BA": (1, 8), "WS_k": (2, 10), "WS_p": (0.01, 0.5)}
 
 
 # ---------------------------------------------------------------------------
-# GIC scoring (apples-to-apples with the pipeline)
+# GIC scoring
 # ---------------------------------------------------------------------------
 
 def gic_of(real_nx, model_name, gen_fn, n_runs, seed):
@@ -164,7 +170,6 @@ def closed_form_params(G):
     k_ws = max(2, int(round(kbar)))
     if k_ws % 2 == 1:
         k_ws += 1
-    # WS rewiring p from clustering moment.
     C_obs = nx.average_clustering(G)
     if k_ws > 2:
         C0 = 3 * (k_ws - 2) / (4 * (k_ws - 1))
@@ -179,7 +184,7 @@ def closed_form_params(G):
 
 
 # ---------------------------------------------------------------------------
-# Current-style grid search (fixed interval, pick min GIC = best case for grid)
+# Grid search (fixed interval, pick min GIC = best case for grid)
 # ---------------------------------------------------------------------------
 
 def grid_best(real_nx, model_name, n, build_gen, lo, hi, n_runs, seed):
@@ -206,10 +211,7 @@ def grid_best_ws(real_nx, n, n_runs, seed):
 
 
 # ---------------------------------------------------------------------------
-# LG scored FAIRLY: burn-in + ensemble-mean spectral density (same convention
-# as the baselines, which average n_runs sample densities). This avoids the
-# pipeline's "best graph over a noisy trajectory" cherry-pick, which is
-# downward-biased and budget-dependent (see diag_lg_convergence.py).
+# LG scored fairly: burn-in + ensemble-mean spectral density
 # ---------------------------------------------------------------------------
 
 def _lg_burn(n):
@@ -228,7 +230,6 @@ def lg_gic_one_d(adj, d, seed):
 
     dens = []
     if d == 0:
-        # d=0 equilibrium is ER(expit(sigma)); ensemble of independent draws.
         for r in range(N_RUNS):
             g = nx.from_numpy_array(_direct_er_at_sigma(n, sigma, seed=seed + r))
             dens.append(scorer.compute_spectral_density(g)[0])
@@ -259,85 +260,98 @@ def lg_gic(adj, seed):
 
 
 # ---------------------------------------------------------------------------
+# Data loading
+# ---------------------------------------------------------------------------
+
+def _load_graphml(path):
+    """Undirected, unweighted (binary) largest connected component."""
+    G = nx.read_graphml(path)
+    G = nx.Graph(G)
+    G.remove_edges_from(nx.selfloop_edges(G))
+    cc = max(nx.connected_components(G), key=len)
+    return nx.convert_node_labels_to_integers(G.subgraph(cc).copy())
+
+
+def _sample_subjects():
+    """(scale, path) pairs: PER_SCALE seeded subjects from each scale present."""
+    rng = random.Random(SEED)
+    chosen = []
+    for scale in SCALES:
+        paths = sorted((_repo_root / "data" / "brain_graph" / scale).glob("*.graphml"))
+        if not paths:
+            print(f"  scale '{scale}' not found under data/brain_graph/ — skipping")
+            continue
+        k = min(PER_SCALE, len(paths))
+        for p in sorted(rng.sample(paths, k)):
+            chosen.append((scale, p))
+    return chosen
+
+
+# ---------------------------------------------------------------------------
 # Main
 # ---------------------------------------------------------------------------
 
-def _ensure_data(data_dir):
-    """Extract data/misc/gplus.tar.gz if the .edges files are not present."""
-    if data_dir.exists() and any(data_dir.glob("*.edges")):
-        return
-    tarball = _repo_root / "data" / "misc" / "gplus.tar.gz"
-    if not tarball.exists():
-        return
-    import tarfile
-    print(f"extracting {tarball.relative_to(_repo_root)} ...")
-    with tarfile.open(tarball) as tf:
-        tf.extractall(_repo_root / "data" / "misc")
-
-
 def main():
-    data_dir = _repo_root / "data" / "misc" / "gplus"
-    _ensure_data(data_dir)
-    files = sorted(data_dir.glob("*.edges"))
-    out_dir = _here / "runs" / "gplus_closedform"
+    out_dir = _here / "runs" / "human_connectomes_closedform"
     out_dir.mkdir(parents=True, exist_ok=True)
 
-    print(f"closed-form baseline experiment  seed={SEED}  quick={QUICK}  "
-          f"window=[{MIN_NODES},{MAX_NODES}]  max_nets={MAX_NETS}  "
-          f"n_runs={N_RUNS}  grid_points={GRID_POINTS}  "
-          f"lg_burn=max({LG_BURN_MIN},{LG_BURN_PER_N}n)  lg_stride=max({LG_STRIDE_MIN},{LG_STRIDE_PER_N}n)")
+    print(f"human-connectomes closed-form experiment  seed={SEED}  quick={QUICK}  "
+          f"scales={SCALES}  per_scale={PER_SCALE}  window=[{MIN_NODES},{MAX_NODES}]  "
+          f"n_runs={N_RUNS}  grid_points={GRID_POINTS}")
+
+    subjects = _sample_subjects()
+    if not subjects:
+        print("No subjects found — brain_graph data is gitignored; place the "
+              "OASIS-3 .graphml scales under data/brain_graph/ first.")
+        return
 
     rows = []
     picked = 0
-    for f in files:
-        if picked >= MAX_NETS:
-            break
-        G = nx.read_edgelist(f, create_using=nx.DiGraph).to_undirected()
-        G = nx.convert_node_labels_to_integers(G)
+    for scale, path in subjects:
+        try:
+            G = _load_graphml(path)
+        except Exception as e:
+            print(f"  skip {path.stem}: {e}")
+            continue
         n = G.number_of_nodes()
         if not (MIN_NODES <= n <= MAX_NODES):
             continue
         picked += 1
         t0 = time.perf_counter()
         cf = closed_form_params(G)
-        adj = nx.to_numpy_array(G)
+        adj = nx.to_numpy_array(G, weight=None)  # binarize
         real_nx = nx.from_numpy_array(adj)
-        name = f.stem[:10]
+        name = f"{scale.replace('oasis3_graphmls_', '').replace('repeated_10_', 'rep_')}/{path.stem[:8]}"
         print(f"\n[{picked}] {name}  n={n}  E={cf['E']}  density={cf['density']:.3f}  "
               f"kbar={cf['kbar']:.1f}  C={cf['C_obs']:.3f}")
 
-        # LG
         lg_val, lg_d = lg_gic(adj, SEED)
         print(f"    LG     gic={lg_val:.3f} (d={lg_d})")
 
-        # ER
         er_grid = grid_best(real_nx, "ER", n,
                             lambda v: _er(n, v), *GRID_INTERVALS["ER"], N_RUNS, SEED)
         er_cf, _ = gic_of(real_nx, "ER", _er(n, cf["ER"]), N_RUNS, SEED)
         print(f"    ER     grid={er_grid[0]:.3f} (p={er_grid[1]:.3f})   "
               f"cf={er_cf:.3f} (p={cf['ER']:.3f})")
 
-        # BA
         ba_grid = grid_best(real_nx, "BA", n,
                             lambda v: _ba(n, v), *GRID_INTERVALS["BA"], N_RUNS, SEED)
         ba_cf, _ = gic_of(real_nx, "BA", _ba(n, cf["BA"]), N_RUNS, SEED)
         print(f"    BA     grid={ba_grid[0]:.3f} (m={ba_grid[1]:.1f})   "
               f"cf={ba_cf:.3f} (m={cf['BA']})")
 
-        # WS
         ws_grid = grid_best_ws(real_nx, n, N_RUNS, SEED)
         ws_cf, _ = gic_of(real_nx, "WS", _ws(n, cf["WS_k"], cf["WS_p"]), N_RUNS, SEED)
         print(f"    WS     grid={ws_grid[0]:.3f} (k,p={ws_grid[1]})   "
               f"cf={ws_cf:.3f} (k={cf['WS_k']},p={cf['WS_p']:.3f})")
 
-        # KR / GRG closed-form only
         kr_cf, _ = gic_of(real_nx, "KR", _kr(n, cf["KR"]), N_RUNS, SEED)
         grg_cf, _ = gic_of(real_nx, "GRG", _grg(n, cf["GRG"]), N_RUNS, SEED)
         print(f"    KR cf={kr_cf:.3f} (d={cf['KR']})   GRG cf={grg_cf:.3f} (r={cf['GRG']:.3f})"
               f"   [{time.perf_counter() - t0:.0f}s]")
 
         rows.append(dict(
-            name=name, n=n, E=cf["E"], density=cf["density"], kbar=cf["kbar"],
+            scale=scale, name=path.stem, n=n, E=cf["E"], density=cf["density"], kbar=cf["kbar"],
             LG=lg_val, LG_d=lg_d,
             ER_grid=er_grid[0], ER_cf=er_cf,
             BA_grid=ba_grid[0], BA_cf=ba_cf,
@@ -347,12 +361,12 @@ def main():
 
     df = pd.DataFrame(rows)
     if df.empty:
-        print("\nNo networks in window.")
+        print("\nNo subjects in window.")
         return
     df.to_csv(out_dir / "results.csv", index=False)
 
     print("\n" + "=" * 78)
-    print("AGGREGATE (mean GIC across networks, lower = better)")
+    print("AGGREGATE (mean GIC across subjects, lower = better)")
     print("=" * 78)
     for fam in ("ER", "BA", "WS"):
         g = df[f"{fam}_grid"].mean()
@@ -363,14 +377,16 @@ def main():
     print(f"  KR(cf)={df['KR_cf'].mean():.3f}   GRG(cf)={df['GRG_cf'].mean():.3f}   "
           f"LG={df['LG'].mean():.3f}")
 
-    # Ranking among LG + closed-form baselines (lower GIC = rank 1)
     rank_cols = {"LG": "LG", "ER": "ER_cf", "BA": "BA_cf", "WS": "WS_cf",
                  "KR": "KR_cf", "GRG": "GRG_cf"}
     ranks = df[list(rank_cols.values())].rank(axis=1)
     ranks.columns = list(rank_cols.keys())
     print("\nMean rank (LG + closed-form baselines, lower = better):")
     print(ranks.mean().sort_values().to_string())
-    print(f"\nWrote {out_dir / 'results.csv'}")
+    print("\nPer-scale mean GIC (LG vs grid baselines):")
+    print(df.groupby("scale")[["n", "density", "LG", "ER_grid", "BA_grid", "WS_grid"]]
+          .mean().round(3).to_string())
+    print(f"\nWrote {out_dir / 'results.csv'}  ({len(df)} subjects)")
 
 
 if __name__ == "__main__":
