+++ scripts/gic/run_lg_twitter_gic.py
@@ -1,30 +1,30 @@
 #!/usr/bin/env python3
-"""Fit LG + ER/WS/BA on every animal connectome and rank them by GIC.
+"""Fit LG + ER/WS/BA on every Twitter SNAP ego network and rank them by GIC.
 
-For each ``data/connectomes/*.graphml`` file, fits the Logit-Graph model
-(AIC d̂ + σ̂) and three baselines (Erdős–Rényi, Watts–Strogatz,
-Barabási–Albert), scores each by spectral GIC against the real graph,
-and ranks them. Outputs per-graph reports under
-``scripts/gic/runs/connectomes/{graph}/`` and aggregate tables/plots.
+For each ``data/misc/twitter/*.edges`` file with ``MIN_NODES ≤ n ≤ MAX_NODES``,
+fits the Logit-Graph model (AIC d̂ + σ̂) and three baselines (Erdős–Rényi,
+Watts–Strogatz, Barabási–Albert), scores each by spectral GIC against the
+real graph, and ranks them. Outputs per-graph reports under
+``scripts/gic/runs/twitter/{graph}/`` and aggregate tables/plots.
 
-Env-var overrides:
-  LG_CONN_MAX_NODES     cap on |V|     (default 2000)
-  LG_CONN_MIN_NODES     floor on |V|   (default 20)
-  LG_CONN_LG_ITER       LG MCMC cap    (default 1500)
-  LG_CONN_GRID_POINTS   baseline grid  (default 3)
-  LG_CONN_N_RUNS        baseline reps  (default 1)
-  LG_CONN_USE_CACHE     reload finished (0/1, default 1)
-  LG_CONN_WORKERS       parallel proc count (default cpu-1)
-  LG_CONN_QUICK         set to 1 for smoke (MAX_NODES=300, fewer iter)
+Env-var overrides (all optional):
+  LG_TWITTER_MAX_NODES     cap on |V| (default 300, set to "none" for no cap)
+  LG_TWITTER_MIN_NODES     floor on |V| (default 50)
+  LG_TWITTER_LG_ITER       LG max_iterations override (default 2000)
+  LG_TWITTER_GRID_POINTS   baseline grid resolution (default 3)
+  LG_TWITTER_N_RUNS        baseline ensemble size (default 1)
+  LG_TWITTER_USE_CACHE     reload finished networks (0/1, default 1)
+  LG_TWITTER_QUICK         set to 1 for smoke (MAX_NODES=150, LG_ITER=1000)
 
-  make gic-connectomes        full preset (~3-5 min on 4 cores)
-  make gic-connectomes-quick  smoke (~30s)
+  make lg-gic-twitter         full preset, ~3-5 min on 4 cores
+  make lg-gic-twitter-quick   smoke run (~30s on 4 cores)
 """
 from __future__ import annotations
 
 import os
 import sys
 import time
+import traceback
 import warnings
 from pathlib import Path
 from typing import Optional
@@ -51,30 +51,8 @@ def _fmt_secs(s: float) -> str:
     return f"{int(m):2d}m{int(rem):02d}s"
 
 
-def _load_graphml_graph(path):
-    """Drop-in replacement for platform_fit_utils.load_edges that handles
-    .graphml. Returns the largest connected component as a relabeled
-    ``nx.Graph`` with self-loops removed."""
-    import networkx as nx
-
-    G = nx.read_graphml(path)
-    G = nx.Graph(G)  # collapse multi-edges + direction
-    G.remove_edges_from(nx.selfloop_edges(G))
-    if G.number_of_nodes() == 0:
-        raise ValueError(f"Empty graph loaded from {path}")
-    if G.number_of_edges() == 0:
-        raise ValueError(f"Edgeless graph loaded from {path}")
-    cc = max(nx.connected_components(G), key=len)
-    return nx.convert_node_labels_to_integers(G.subgraph(cc).copy())
-
-
-def _patch_load_edges():
-    import platform_fit_utils as pfu
-
-    pfu.load_edges = _load_graphml_graph
-
-
 def main() -> None:
+    # Unbuffered output so progress shows up live
     sys.stdout.reconfigure(line_buffering=True)
     for v in ("OPENBLAS_NUM_THREADS", "OMP_NUM_THREADS", "MKL_NUM_THREADS"):
         os.environ.setdefault(v, "1")
@@ -88,8 +66,6 @@ def main() -> None:
     if str(_here) not in sys.path:
         sys.path.insert(0, str(_here))
 
-    _patch_load_edges()
-
     import platform_fit_utils as pfu  # noqa: E402
     from platform_fit_utils import (  # noqa: E402
         PlatformConfig,
@@ -101,17 +77,17 @@ def main() -> None:
     )
     import pandas as pd  # noqa: E402
 
-    quick = os.environ.get("LG_CONN_QUICK", "0") == "1"
-    max_nodes_default = 300 if quick else 2000
-    lg_iter_default = 800 if quick else 1500
+    quick = os.environ.get("LG_TWITTER_QUICK", "0") == "1"
+    max_nodes_default = 150 if quick else 300
+    lg_iter_default = 1000 if quick else 2000
 
-    max_nodes = _get_optional_int("LG_CONN_MAX_NODES", max_nodes_default)
-    min_nodes = _get_int("LG_CONN_MIN_NODES", 20)
-    lg_iter_cap = _get_int("LG_CONN_LG_ITER", lg_iter_default)
-    grid_points = _get_int("LG_CONN_GRID_POINTS", 3)
-    n_runs = _get_int("LG_CONN_N_RUNS", 1)
-    use_cache = os.environ.get("LG_CONN_USE_CACHE", "1") == "1"
-    workers = _get_int("LG_CONN_WORKERS", max(1, (os.cpu_count() or 2) - 1))
+    max_nodes = _get_optional_int("LG_TWITTER_MAX_NODES", max_nodes_default)
+    min_nodes = _get_int("LG_TWITTER_MIN_NODES", 50)
+    lg_iter_cap = _get_int("LG_TWITTER_LG_ITER", lg_iter_default)
+    grid_points = _get_int("LG_TWITTER_GRID_POINTS", 3)
+    n_runs = _get_int("LG_TWITTER_N_RUNS", 1)
+    use_cache = os.environ.get("LG_TWITTER_USE_CACHE", "1") == "1"
+    workers = _get_int("LG_TWITTER_WORKERS", max(1, (os.cpu_count() or 2) - 1))
 
     _original_lg_max = pfu.lg_max_iterations
 
@@ -121,8 +97,8 @@ def main() -> None:
     pfu.lg_max_iterations = _capped
 
     cfg = PlatformConfig(
-        platform="connectomes",
-        glob_pattern="connectomes/*.graphml",
+        platform="twitter",
+        glob_pattern="misc/twitter/*.edges",
         min_nodes=min_nodes,
         max_nodes=max_nodes,
         other_model_n_runs=n_runs,
@@ -133,25 +109,27 @@ def main() -> None:
         run_dir=_here / "runs",
     )
 
-    print(
-        f"connectomes GIC ranking  min_nodes={min_nodes}  max_nodes={max_nodes}  "
+    banner = (
+        f"twitter GIC ranking  min_nodes={min_nodes}  max_nodes={max_nodes}  "
         f"lg_iter≤{lg_iter_cap}  grid_points={grid_points}  n_runs={n_runs}  "
-        f"cache={use_cache}  workers={workers}  quick={quick}"
+        f"cache={use_cache}  quick={quick}"
     )
+    print(banner)
 
-    graph_files, sizes_by_stem = _discover(cfg)
+    graph_files, sizes_by_stem = _fast_discover(cfg)
     if not graph_files:
-        print("No connectomes matched the size window; nothing to do.")
+        print("No graphs matched the size window; nothing to do.")
         return
 
     sizes = [(stem, sizes_by_stem[stem]) for stem in (p.stem for p in graph_files)]
     avg_n = sum(n for _, n in sizes) / len(sizes)
     print(
-        f"\nDiscovered {len(graph_files)} connectomes  "
+        f"\nDiscovered {len(graph_files)} networks  "
         f"|V|: min={min(n for _, n in sizes)}  max={max(n for _, n in sizes)}  "
         f"mean={avg_n:.0f}"
     )
 
+    # Pre-check the cache to estimate how many fits are actually needed
     pending = []
     cached_count = 0
     for p in graph_files:
@@ -160,13 +138,19 @@ def main() -> None:
             cached_count += 1
         else:
             pending.append(p)
+    eta_per_fit = max(5.0, 0.05 * avg_n) if pending else 0  # ~5s+ per graph
     print(
-        f"Cache: {cached_count}/{len(graph_files)} hits, {len(pending)} fits pending\n"
+        f"Cache: {cached_count}/{len(graph_files)} hits, {len(pending)} fits pending "
+        f"(rough ETA per fit ≈ {_fmt_secs(eta_per_fit)})"
+    )
+    print(
+        f"Total wall-time estimate: {_fmt_secs(len(pending) * eta_per_fit)} "
+        f"(serial; less with adaptive)\n"
     )
 
     logger = setup_platform_logging(cfg.run_dir)
     logger.info(
-        "=== connectomes GIC sweep  (%d networks, %d cached, %d to fit) ===",
+        "=== twitter GIC sweep  (%d networks, %d cached, %d to fit) ===",
         len(graph_files), cached_count, len(pending),
     )
 
@@ -175,6 +159,7 @@ def main() -> None:
     failures = []
     t_start = time.perf_counter()
 
+    # Serial pass: replay cached cells
     pending_with_idx: list[tuple[int, Path]] = []
     for i, edge_path in enumerate(graph_files, start=1):
         graph_name = edge_path.stem
@@ -193,17 +178,18 @@ def main() -> None:
                 continue
         pending_with_idx.append((i, edge_path))
 
+    # Parallel pass: fresh fits across workers
     if pending_with_idx:
         worker_count = max(1, min(workers, len(pending_with_idx)))
         print(
-            f"\nLaunching {worker_count} worker(s) for "
+            f"\nLaunching {worker_count} worker process(es) for "
             f"{len(pending_with_idx)} fresh fits ...\n"
         )
         if worker_count == 1:
-            results_iter = [
+            results_iter = (
                 _fit_worker((str(p), _cfg_to_dict(cfg), lg_iter_cap, i, len(graph_files)))
                 for i, p in pending_with_idx
-            ]
+            )
         else:
             from concurrent.futures import ProcessPoolExecutor, as_completed
             tasks = [
@@ -269,6 +255,7 @@ def main() -> None:
 
 
 def _cfg_to_dict(cfg) -> dict:
+    """Serialize PlatformConfig for transport across process boundary."""
     return {
         "platform": cfg.platform,
         "glob_pattern": cfg.glob_pattern,
@@ -281,13 +268,17 @@ def _cfg_to_dict(cfg) -> dict:
         "display_plots": cfg.display_plots,
         "use_cache": cfg.use_cache,
         "data_root": str(cfg.data_root),
+        # Pass parent of run_dir because PlatformConfig.__post_init__ appends platform
         "run_dir_parent": str(cfg.run_dir.parent),
     }
 
 
 def _fit_worker(args):
-    """Runs in a child process. Writes per-graph artifacts to disk; returns
-    only meta + summary across the IPC boundary."""
+    """Run in a child process. Returns (status, name, payload).
+
+    Worker writes all per-graph artifacts to disk (via fit_one_network);
+    only the small meta dict + summary DataFrame are shipped back over IPC.
+    """
     edge_path_str, cfg_dict, lg_iter_cap, i, total = args
     import sys as _sys
     import os as _os
@@ -302,7 +293,6 @@ def _fit_worker(args):
         if _p not in _sys.path:
             _sys.path.insert(0, _p)
 
-    _patch_load_edges()
     import platform_fit_utils as pfu
     from platform_fit_utils import PlatformConfig, fit_one_network, setup_platform_logging
 
@@ -313,7 +303,11 @@ def _fit_worker(args):
     cfg = PlatformConfig(**cfg_kwargs)
 
     _orig_lg_max = pfu.lg_max_iterations
-    pfu.lg_max_iterations = lambda n: min(_orig_lg_max(n), lg_iter_cap)
+
+    def _capped(n: int) -> int:
+        return min(_orig_lg_max(n), lg_iter_cap)
+
+    pfu.lg_max_iterations = _capped
 
     edge_path = _Path(edge_path_str)
     logger = setup_platform_logging(cfg.run_dir, name=f"worker_{_os.getpid()}")
@@ -325,27 +319,50 @@ def _fit_worker(args):
         return ("err", edge_path.stem, f"{exc}\n{_tb.format_exc()}")
 
 
-def _discover(cfg):
-    """Load each .graphml once and filter by node-count bounds.
+def _peek_size(edge_path: Path) -> int:
+    """Get |V| of an .edges file without networkx parsing."""
+    nodes = set()
+    with open(edge_path, "r", encoding="utf-8") as f:
+        for line in f:
+            parts = line.split()
+            if len(parts) >= 2:
+                nodes.add(parts[0])
+                nodes.add(parts[1])
+    return len(nodes)
 
-    Connectome files are graphml so we can't byte-prefilter as we did for
-    gplus's plain edge lists; but there are only ~20 of them and they load
-    in <2s total.
+
+def _fast_discover(cfg):
+    """File-size pre-filter + cheap node-count peek, skipping networkx parsing.
+
+    The reference ``platform_fit_utils.discover_graph_files`` loads every
+    candidate via ``nx.read_edgelist`` to count nodes — for the twitter
+    collection this scans ~130 files (incl. multi-MB ones above MAX_NODES)
+    and takes minutes. Here we (1) drop files whose byte size implies
+    n > 2·max_nodes (lots of slack) and (2) count unique node tokens
+    in a single linear pass for the survivors.
     """
     paths = sorted(cfg.data_root.glob(cfg.glob_pattern))
-    paths = [p for p in paths if p.suffix == ".graphml" and p.is_file()]
+    paths = [p for p in paths if p.suffix == ".edges" and p.is_file()]
     sizes: dict[str, int] = {}
     kept: list[tuple[Path, int]] = []
+    if cfg.max_nodes is not None:
+        # Empirically each edge in this dataset is ~20 bytes; n*(n-1)/2 edges
+        # at upper bound → safe size threshold = 50 · n²·2 = 100 · n² bytes.
+        max_bytes = 100 * cfg.max_nodes * cfg.max_nodes
+    else:
+        max_bytes = None
     for p in paths:
-        try:
-            G = _load_graphml_graph(p)
-        except (ValueError, OSError) as exc:
-            print(f"  skipped {p.name}: {exc}")
+        if max_bytes is not None and p.stat().st_size > max_bytes:
             continue
-        n = G.number_of_nodes()
+        try:
+            n = _peek_size(p)
+        except OSError:
+            continue
         if n < cfg.min_nodes:
             continue
         if cfg.max_nodes is not None and n > cfg.max_nodes:
+            continue
+        if n == 0:
             continue
         sizes[p.stem] = n
         kept.append((p, n))
